requirements.txt
openai
gradio
python-dotenv
langchain
langchain-openai
langchain-chroma
matplotlib
scikit-learn
numpy
plotly